Support `SHF_GNU_RETAIN`
This section flag (`R`) is used for `#[used(linker)]` in Rust and
`__attribute__((retain))` in GCC. It marks a section as being a GC
root to avoid removing it with `--gc-sections`.

This parses the flag, forwards it through the linker, and then uses it
in the GC pass. To make it work properly, it also accepting the
GNU ABI, since SHF_GNU_RETAIN is a GNU extension

To make it match BFD as closely as possible, I moved section merging before gc-sections.
With the following assembly:
```asm
.section .rodata.retained_same_name,"a",%progbits
    .asciz "I also happen to be retained because I have the same name"

.section .rodata.retained_same_name,"aR",%progbits
    .asciz "I really want to be retained"
```

BFD would produce a single section that contained *both* contents, being
evidence that the section were merged first (preserving `R` on the
result) and then being gced. This move didn't lead to any other test
failures.